chore: mc app context selection
