+++ libs/mechanicalcompletionapp/src/lib/config/workspaceConfig.tsx
@@ -8,12 +8,11 @@ import { useCallback } from 'react';
 import { contextConfig } from './contextConfig';
 import { responseParser } from './responseConfig';
 
-const useContextId = () => {
-  return '2d489afd-d3ec-43f8-b7ca-cf2de5f39a89';
+type WorkspaceWrapperProps = {
+  contextId: string;
 };
-export const WorkspaceWrapper = () => {
+export const WorkspaceWrapper = ({ contextId }: WorkspaceWrapperProps) => {
   const dataProxy = useHttpClient('data-proxy');
-  const contextId = useContextId();
   const getResponseAsync = useCallback(async () => {
     const mcpkgs = await dataProxy.fetch(`/api/contexts/${contextId}/mc-pkgs`);
     return mcpkgs;
